+++ src/main/java/fr/vinvin129/budgetmanager/ihm/views/controllers/create/category/CreateStandardCategoryController.java
@@ -36,7 +36,7 @@ public class CreateStandardCategoryController implements CreateCategory {
             throw new CreateCategoryException("le nom et l'allocation de la catégorie ne doivent pas être vide");
         }
         try {
-            return new StandardCategory(name, Integer.parseInt(allocation));
+            return new StandardCategory(name, Double.parseDouble(allocation));
         } catch (NumberFormatException e) {
             throw new CreateCategoryException("la valeur du champ allocation doit être un nombre");
         } catch (IllegalCategorySizeException e) {
